Fix turn lanes UI

+++ Sources/Controllers/Map/Widgets/OALanesDrawable.h
@@ -16,8 +16,8 @@
 @property (nonatomic) BOOL leftSide;
 @property (nonatomic) BOOL imminent;
 
-@property (nonatomic, readonly) int height;
-@property (nonatomic, readonly) int width;
+@property (nonatomic, readonly) CGFloat height;
+@property (nonatomic, readonly) CGFloat width;
 
 - (instancetype) initWithScaleCoefficient:(float)scaleCoefficient;
